фиксы по замечаниям
