+++ src/mock/event.js
@@ -1,15 +1,6 @@
 import {EventTypes} from '../const.js';
-import {castTimeFormat} from '../utils.js';
+import {castTimeFormat, getRandomArrayItem, getRandomIntegerNumber, getRandomDate} from '../utils.js';
 
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length);
-
-  return array[randomIndex];
-};
-
-const getRandomIntegerNumber = (min, max) => {
-  return min + Math.floor(max * Math.random());
-};
 
 const generateOptions = (qty) => {
   const randomOffers = [];
@@ -44,16 +35,6 @@ const generatePictures = (qty) => {
     });
   }
   return pictures;
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomIntegerNumber(0, 7);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
 };
 
 const DescriptionItems = [
